Save a couple of bytes
... by using fewer bytes.

+++ database/books.cpp
@@ -184,7 +184,7 @@ string get_onlinebible_from_id (book_id id)
 }
 
 
-int get_order_from_id (book_id id)
+short get_order_from_id (book_id id)
 {
   for (unsigned int i = 0; i < data_count; i++) {
     if (id == books_table[i].id) {
